object型に変換

+++ src/TypeSafe/Main.java
@@ -7,7 +7,7 @@ public class Main {
 
 	public static void prints(Object a, Object b) {
 
-		for(int i = 0; i<b; i++) {
+		for(int i = 0; i<(Integer)b; i++) {
 			System.out.println(a);
 		}
